tests/wpt/web-platform-tests/css/css-multicol/parsing/column-gap-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Multi-column Layout: getComputedValue().columnGap</title>
<link rel="help" href="https://drafts.csswg.org/css-multicol/#propdef-column-gap">
<meta name="assert" content="column-gap computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<style>
  #target {
    font-size: 40px;
  }
</style>
<script>
test_computed_value("column-gap", "normal");

test_computed_value("column-gap", "calc(10px + 0.5em)", "30px");
test_computed_value("column-gap", "calc(10px - 0.5em)", "0px");
test_computed_value("column-gap", "40%");
test_computed_value("column-gap", "calc(50% + 60px)");
</script>
</body>
</html>
